20230522 값이 왜 안 넘어감
20230522 값이 왜 안 넘어감

+++ src/main/java/com/project/entity/Board.java
@@ -1,5 +1,6 @@
 package com.project.entity;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -31,7 +32,7 @@ public class Board {
     
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "SEQ_BOARD_NO")
-    private Long no;
+    private BigInteger no;
 
     private String title;
